+++ LocalPayment/src/main/java/com/braintreepayments/api/localpayment/LocalPaymentAuthResult.kt
@@ -1,5 +1,7 @@
 package com.braintreepayments.api.localpayment
 
+import com.braintreepayments.api.BrowserSwitchFinalResult
+
 /**
  * Result of the local payment web flow received from [LocalPaymentLauncher.handleReturnToAppFromBrowser].
  */
@@ -8,7 +10,7 @@ sealed class LocalPaymentAuthResult {
     /**
      * A successful result that should be passed to [LocalPaymentClient.tokenize] to complete the flow
      */
-    class Success(internal val paymentAuthInfo: LocalPaymentAuthResultInfo) : LocalPaymentAuthResult()
+    class Success(internal val browserSwitchSuccess: BrowserSwitchFinalResult.Success) : LocalPaymentAuthResult()
 
     /**
      * The browser switch failed.